Refactor: 공통 예외 이름 변경, 위치 변경
- CommonException -> TissueException
- 도메인 관련 예외를 common.exception.domain 아래로 이동
