+++ src/main/java/com/uranus/taskmanager/api/common/exception/TissueException.java
@@ -5,15 +5,15 @@ import org.springframework.http.HttpStatus;
 import lombok.Getter;
 
 @Getter
-public class CommonException extends RuntimeException {
+public class TissueException extends RuntimeException {
 	private final HttpStatus httpStatus;
 
-	public CommonException(String message, HttpStatus httpStatus) {
+	public TissueException(String message, HttpStatus httpStatus) {
 		super(message);
 		this.httpStatus = httpStatus;
 	}
 
-	public CommonException(String message,
+	public TissueException(String message,
 		HttpStatus httpStatus, Throwable cause) {
 		super(message, cause);
 		this.httpStatus = httpStatus;
