+++ src/main/java/com/uranus/taskmanager/api/common/exception/domain/InvitationException.java
@@ -1,8 +1,10 @@
-package com.uranus.taskmanager.api.common.exception;
+package com.uranus.taskmanager.api.common.exception.domain;
 
 import org.springframework.http.HttpStatus;
 
-public class InvitationException extends CommonException {
+import com.uranus.taskmanager.api.common.exception.TissueException;
+
+public class InvitationException extends TissueException {
 
 	public InvitationException(String message, HttpStatus httpStatus) {
 		super(message, httpStatus);
