+++ src/main/java/com/uranus/taskmanager/api/common/exception/domain/MemberException.java
@@ -1,8 +1,10 @@
-package com.uranus.taskmanager.api.common.exception;
+package com.uranus.taskmanager.api.common.exception.domain;
 
 import org.springframework.http.HttpStatus;
 
-public abstract class MemberException extends CommonException {
+import com.uranus.taskmanager.api.common.exception.TissueException;
+
+public abstract class MemberException extends TissueException {
 
 	public MemberException(String message, HttpStatus httpStatus) {
 		super(message, httpStatus);
