+++ src/main/java/com/uranus/taskmanager/api/common/exception/domain/WorkspaceException.java
@@ -1,8 +1,10 @@
-package com.uranus.taskmanager.api.common.exception;
+package com.uranus.taskmanager.api.common.exception.domain;
 
 import org.springframework.http.HttpStatus;
 
-public abstract class WorkspaceException extends CommonException {
+import com.uranus.taskmanager.api.common.exception.TissueException;
+
+public abstract class WorkspaceException extends TissueException {
 	public WorkspaceException(String message, HttpStatus httpStatus) {
 		super(message, httpStatus);
 	}
