+++ src/main/java/com/uranus/taskmanager/api/invitation/exception/InvalidInvitationStatusException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.invitation.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.InvitationException;
+import com.uranus.taskmanager.api.common.exception.domain.InvitationException;
 
 public class InvalidInvitationStatusException extends InvitationException {
 	private static final String MESSAGE = "Invitation has already been processed";
