+++ src/main/java/com/uranus/taskmanager/api/invitation/exception/InvitationAlreadyExistsException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.invitation.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.InvitationException;
+import com.uranus.taskmanager.api.common.exception.domain.InvitationException;
 
 public class InvitationAlreadyExistsException extends InvitationException {
 	private static final String MESSAGE = "An invitation for this member is already pending";
