+++ src/main/java/com/uranus/taskmanager/api/member/exception/OwnedWorkspaceExistsException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.member.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.MemberException;
+import com.uranus.taskmanager.api.common.exception.domain.MemberException;
 
 public class OwnedWorkspaceExistsException extends MemberException {
 	private static final String MESSAGE = "You currently have one or more owned workspaces.";
