+++ src/main/java/com/uranus/taskmanager/api/position/exception/DuplicatePositionNameException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.position.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.PositionException;
+import com.uranus.taskmanager.api.common.exception.domain.PositionException;
 
 public class DuplicatePositionNameException extends PositionException {
 	private static final String MESSAGE = "The name for the position is duplicate in this workspace.";
