+++ src/main/java/com/uranus/taskmanager/api/position/exception/PositionInUseException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.position.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.PositionException;
+import com.uranus.taskmanager.api.common.exception.domain.PositionException;
 
 public class PositionInUseException extends PositionException {
 	private static final String MESSAGE = "Cannot delete position that is in use.";
