+++ src/main/java/com/uranus/taskmanager/api/position/exception/PositionNotFoundException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.position.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.PositionException;
+import com.uranus.taskmanager.api.common.exception.domain.PositionException;
 
 public class PositionNotFoundException extends PositionException {
 	private static final String MESSAGE = "Could not find the following position for the workspace.";
