+++ src/main/java/com/uranus/taskmanager/api/security/authentication/exception/InvalidLoginIdentityException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.security.authentication.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.AuthenticationException;
+import com.uranus.taskmanager.api.common.exception.domain.AuthenticationException;
 
 public class InvalidLoginIdentityException extends AuthenticationException {
