+++ src/main/java/com/uranus/taskmanager/api/security/authorization/exception/InvalidWorkspaceCodeInUriException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.security.authorization.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.AuthorizationException;
+import com.uranus.taskmanager.api.common.exception.domain.AuthorizationException;
 
 public class InvalidWorkspaceCodeInUriException extends AuthorizationException {
 	private static final String MESSAGE = "The workspace code in the URI is invalid";
