+++ src/main/java/com/uranus/taskmanager/api/security/authorization/exception/UpdatePermissionException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.security.authorization.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.AuthorizationException;
+import com.uranus.taskmanager.api.common.exception.domain.AuthorizationException;
 
 public class UpdatePermissionException extends AuthorizationException {
 	private static final String MESSAGE = "You do not have authorization for update or the authorization has expired";
