+++ src/main/java/com/uranus/taskmanager/api/workspacemember/exception/InvalidWorkspaceCountException.java
@@ -2,7 +2,7 @@ package com.uranus.taskmanager.api.workspacemember.exception;
 
 import org.springframework.http.HttpStatus;
 
-import com.uranus.taskmanager.api.common.exception.WorkspaceMemberException;
+import com.uranus.taskmanager.api.common.exception.domain.WorkspaceMemberException;
 
 public class InvalidWorkspaceCountException extends WorkspaceMemberException {
 	private static final String MESSAGE = "The count for owned workspaces cannot go below zero.";
